Fix tramp-connection-properties in xps-custom
In the update from my old fileset to the new, I garbled the setting of
the tramp-connection-properties for xps-custom.  What should have been
done in 2 calls to add-to-list was accidentally merged into 1, so the
settings for the virtual machine "beastie" were not taking effect.

I could probably find another way to do this, but what I've got works
for now.