Refactor cwmars-deploy-db-updates into cwmars-deploy
Make cwmars-deploy-db-updates into a more generic command that accepts
an argument to be able to deploy any code repository that has a remote
path variable defined.

Update the existing key binding to deploy the db-updates and add key
bindings for the apps, apps_dev, and utilities repositories.

The remote server name is hardcoded for the main DB server hosted by
MOBIUS.