Refactor backup-bash-aliases into my-dotfile-backup-name
This commit makes the function more generic so that it can be used to
backup any dotfile (any file, really) to a particular directory.

The get-backup-version inner function is moved up to a top level
function and renamed to my-get-backup-version.  The top level function
takes a directory and filename argument to search for backup files
matching the pattern directory/filename.[0-9]+ to determine the next
backup revision number to use.

The my-bash-aliases-backup-dir custom variable is renamed to
my-dotfile-backup-dir.

The my-dotfile-backup-name function now uses the argument to determine
the backup filename.  It strips off any leading dots (.) and returns a
full path to the backup file in my-dotfile-backup-dir with the next
version number tacked on the end.  There is no file collision
detection for similar named files from different directories.  If you
need that, you'll need a different method or you will need to expand
this code.

All of this replaces something that I have doing by hand for a few
years with my .bash_aliases file.  I don't change it that often over
time, but I've recently made changes frequently enough that I decided
to automate backup file creation for it.

I have not set up a Git repository for all of my dotfiles.  I may do
that some day, but in the meantime this will do.