Use git worktree on ILS-MGR for Evergreen and OpenSRF
I started using worktrees with Evergreen and OpenSRF on the ILS-MGR
laptop.  To make this cleaner to manage in my source code directory, I
moved the main Evergreen and OpenSRF worktrees from ~/src/Evergreen/
and ~/src/OpenSRF/ to ~/src/Evergreen/main/ and ~/src/OpenSRF/main/
respectively.

If this works out, I may make the same change for other source
projects and/or on my other laptops.