Add another Text::CSV skeleton
This one, text-csv-readloop, is used to add a read loop to read from a
CSV that has been set up in my usual way.

It has a marker and an interesting place, so it can be wrapped around
existing code. The latter allows you to fill in the body of the loop
and then wrap the while block around it.