Add cwmars-copy-rows skeleton
I often need to copy new rows from production to a test database in
order to test some database update scripts.  This skeleton writes out
a file that can be run through psql to copy data from production and
outputs files that can be used to copy the data into the test
database.