Update ils-mg-custom.el
I added xml-rpc in a vain attempt to use dokuwiki.el. I'm going to
leave it for a while in case we update the configuration of the
community wiki, or in case I have another use for it.

It also appears that the inferior-lisp-program setting changed,
probably after i dropped loading slime helper from QuickLisp.