Improve cwmars-copy-rows skeleton
Now that I have used it, I see some room for improvement in the
cwmars-copy-rows skeleton.  It is more flexible and will do "smart"
things if there is only 1 value in the key values list or the first
entry in the key values list is a comparison operator: ('=', '<', ">',
'<=', ">=').

I also patched a bug with the code to update the sequence in the
destination database.