Use character in group-number arguments
Modify my copy of the group-number function to accept Emacs character
syntax for the separator character.  This is turned into a string for
concatenation.

I tried using this option and because the help says "CHAR" I didn't
realize it needed a string.  I decided to change it to require a
character.