Add punctuate-number utility function
Implement a more generic punctuate-number function to punctuate
numbers into an arbitrary number of arbitrary width fields.  It works
on digits and letters.  This is not because of calc mode supporting up
to base 36.  This was done to have it work with product serial numbers
that might also include ASCII letters.

The group-number function has been refactored to use the
punctuate-number function.

The punctuate-number function was partly inspired the C++ numpunct
template class: https://en.cppreference.com/w/cpp/locale/numpunct.