Fix off by one bug in punctuate-number
We need to subtact 1 from the calculated size in the get-size function
when the modulus of the length of the string by the width of a field
is zero to make up for the fact that we're inserting 1 less
punctuation character.  Otherwise, we get an extrac punctuation at the
beginning of the output string.