Fix the actual bugs in punctuate-number
This commit fixes the get-size local function for real by getting the
check for the modulus being greater than zero correct.  It also no
longer subtracts 1 when the modulus is zero.

When I adapted a prototype of punctuate-number to accept the single
member list, I assumed that butlast was going to be run on the new
list that got created.  In all of my wrangling with the code this
morning that got lost, which created various states of brokenness that
worked under certain conditions.

Part of this fix moves the butlast so that it does affect the list
created when a single member list is passed in.