Modify punctuate-number and get-size one last time
After playing with the function some more, I decided not to do butlast
on the list generated if the sizes list contains 1 member.  The
get-size local function has been modified to return the size for a
list one element shorter.

There is no technical reason for making this change.  My intuition
agrees more with this implementation.