Refine string-match-p criteria for dot file copies
Modify the string used with string-match-p in my after-save-hook for
Bash dotfiles so that it only takes effect when editing such a file in
my home directory.  I was editing on a remote server, or a copy for
the VMs in my CWMARS files, and that version got copied to my backup
directory and overwrote my previous backup.  I only want to make
copies of the file if it is in my local $HOME.