Fix frame sizes on ILS-MGR
I upgraded ILS-MGR to Ubuntu 24.04 again, and GNU Emacs 29.3 still has
issues with frame sizes in Gnome.  This commit adds back a variation
of code that I used during my abortive upgrade trial last fall.

(Since the MIPI camera is not working on the laptop regardless of the
Ubuntu release, I've decided to stick with 24.04 for now.  I've tried
the suggested kernels with the drivers, but they fail to boot the
laptop for various reasons.)