Refactor code to resize frames on ILS-MGR
Refactor the lambda into a named function with defun so that I can
modify it during runtime without having to quit Emacs.

Add code to handle sizing "Ediff" frames to 40x1, because 80x36 is
ridiculous for these windows.  It was this change that made me decide
to refactor to a named function.