Remove mail.sigio.com from tramp-connection-properties
The mail.sigio.com server has been shut down, so I no longer need the
tramp-connection-properties settings to use bash as the shell.

In the cases of ils-mgr and needle this outright removes the
tramp-connection-properties.  In the case of xps the settings not
necessarily related to mail.sigio.com remain.